Date: Sat, 10 Jul 2004 20:39:40 +0200 From: "Philippe Plantier (ayin)" Subject: [SDL] Memory corruption in SDL_ttf
The code, which, in SDL_ttf, handles the copy of glyph data to the SDL surface
has an issue: some glyphs may specify a negative offset, and this is not checked
for. It may cause some nasty memory corruption / crashes.

--HG--
extra : convert_revision : svn%3Ac70aab31-4412-0410-b14c-859654838e24/trunk%401401