Fixed bug: SDL_ttf incorrectly identifies some utf8 characters as "overlong".
Tommy

I noticed that the UTF8_getch() function in SDL_ttf.c is incorrectly identifying some valid UTF-8 characters as "overlong".

The particular character that brought this to my attention was '??' (0x1033a), with the four-byte UTF-8 representation 0xF0908CBA.

The problem is that it only checks if the first byte has no high bits, not if the first 5 meaningful bits are 0. In the extreme 6-byte case, there is only one meaningful bit in the first byte. So fully half of all theoretically valid 6-byte characters are incorrectly identified as overlong.